separate complicated logic in App.tsx
