+++ client/src/hooks/useSettingsPersistence.ts
@@ -1,13 +1,25 @@
 import { useEffect, useState } from "react";
 import { useStore } from "@/core";
+import { useSettingsStore } from "@/core/state/slices/settingsStore";
 
 const SETTINGS_KEY = "reprod.settings";
 
+/**
+ * Handles settings initialization and persistence.
+ *
+ * This hook manages:
+ * - Loading local settings from localStorage on mount
+ * - Saving local settings to localStorage when they change
+ * - Fetching server-side API provider settings on mount
+ */
 export function useSettingsPersistence(): void {
 	const settings = useStore((state) => state.settings);
 	const updateSettings = useStore((state) => state.updateSettings);
 	const [hydrated, setHydrated] = useState(false);
 
+	const { fetchSettings } = useSettingsStore();
+
+	// Load local settings from localStorage
 	useEffect(() => {
 		if (typeof window === "undefined") return;
 
@@ -25,6 +37,7 @@ export function useSettingsPersistence(): void {
 		setHydrated(true);
 	}, [updateSettings]);
 
+	// Save local settings to localStorage
 	useEffect(() => {
 		if (!hydrated || typeof window === "undefined") {
 			return;
@@ -37,4 +50,9 @@ export function useSettingsPersistence(): void {
 			// User experience: Settings changes may not persist, but app remains functional.
 		}
 	}, [hydrated, settings]);
+
+	// Fetch server-side API provider settings
+	useEffect(() => {
+		void fetchSettings();
+	}, [fetchSettings]);
 }
